Add bin2str function.
New UDF converts BINARY values to STRING values.
BINARY typed values now convert to STRING correctly and toString()
formatting is correct throughout.

Timestamp fixed to match schema for actual Flume Event timestamps.

Bugfix improper instantiation of BINARY as byte[] (instead of ByteBuffer).

+++ src/main/java/com/odiago/flumebase/lang/Timestamp.java
@@ -28,7 +28,7 @@ public class Timestamp extends TimestampBase {
     this(millis, 0);
   }
 
-  public Timestamp(long millis, int nanos) {
+  public Timestamp(long millis, long nanos) {
     this.milliseconds = millis;
     this.nanos = nanos;
   }
@@ -40,7 +40,7 @@ public class Timestamp extends TimestampBase {
 
   @Override
   public int hashCode() {
-    return ((int) (this.milliseconds & 0xFFFFFFFF)) ^ (this.nanos * 37);
+    return ((int) (this.milliseconds & 0xFFFFFFFF)) ^ ((int) this.nanos * 37);
   }
 
   @Override
